requirements.txt
Flask==2.2.3
MarkupSafe==2.1.2
Werkzeug==2.2.3
itsdangerous==2.1.2
psutil==5.8.0
plotly==5.5.0
tenacity==8.0.1
boto3==1.9.148
kubernetes==10.0.1
azure-mgmt-containerregistry==2.0.0  # Replace with the latest version
azure-identity==1.6.0  # Replace with the latest version

